Youtube-videos-summarization/requirements.txt
streamlit
google-generativeai
youtube_transcript_api
#-e .
